# logging: do filter last, for improved performance