#### What this PR does

Previously the code would find out what line it was on and format the time, before deciding that the whole log line should be filtered out. Moving the filter to the end avoids most of this work.

Add a benchmark to demonstrate, and a test to validate that the filtering still works.

Very similar to https://github.com/weaveworks/common/pull/227

```
name        old time/op    new time/op    delta
DebugLog-4    2.05µs ±10%    0.21µs ±16%  -89.52%  (p=0.008 n=5+5)

name        old alloc/op   new alloc/op   delta
DebugLog-4      520B ± 0%       96B ± 0%  -81.54%  (p=0.008 n=5+5)

name        old allocs/op  new allocs/op  delta
DebugLog-4      7.00 ± 0%      2.00 ± 0%  -71.43%  (p=0.008 n=5+5)
```

#### Checklist

- [x] Tests updated
- NA Documentation added
- [x] `CHANGELOG.md` updated.